TST: add property-based test for pyssg
* this drafts a first property-based `hypothesis` test
for `py-ssg`, focused on testing the string used for the
group name when creating groups with ssg

* the good news is that I've found no evidence that strange
strings cause issues

* the bad news is that py-ssg/py-margo engine/group creation
contexts appear to have a global state that prevents launching
many of them from the same Python process? This is a problem for
this kind of `hypothesis` case testing. While we can do parametrized
tests with `pytest --fork` to separate the processes, this does not
work with `hypothesis` case generation, nor does the use of the
`pytest-subtests` appoach; open to suggestions here..

* added brief documentation on running the tests